+++ src/components/modules/api/blocks.ts
@@ -87,13 +87,9 @@ export default class BlocksAPI extends Module {
     }
 
     /**
-     * In case of deletion first block we need to set caret to the current Block
+     * After Block deletion currentBlock is updated
      */
-    if (this.Editor.BlockManager.currentBlockIndex === 0) {
-      this.Editor.Caret.setToBlock(this.Editor.BlockManager.currentBlock);
-    } else {
-      this.Editor.Caret.navigatePrevious(true);
-    }
+    this.Editor.Caret.setToBlock(this.Editor.BlockManager.currentBlock, this.Editor.Caret.positions.END);
 
     this.Editor.Toolbar.close();
   }