+++ docs/CHANGELOG.md
@@ -16,6 +16,7 @@
 — `Fix` — Fixed History Back on block deletion by Backspace in Firefox [#967](https://github.com/codex-team/editor.js/pull/967)
 - `Fix` — Fixed `getRangeCount` call if range count is 0 [#938](https://github.com/codex-team/editor.js/issues/938)
 - `New` — Log levels now available to suppress Editor.js console messages [#962](https://github.com/codex-team/editor.js/issues/962) 
+- `Fix` — Fixed wrong navigation on block deletion
 
 ### 2.15.1